Enhance seed_db function
Added a new item to the seed data in the seed_db function. The new item has the name 'Item4' and the description 'This is item 4'. Also, improved the seed_db function to accept a 'delete' argument. If 'delete' is True, the function will delete all existing users and items before seeding the database.